requirements.txt
openai
pydantic
python-dotenv
aiohttp
chromadb
sqlalchemy
pytest
pytest-asyncio
